+++ 20250407/tempCodeRunnerFile.c
@@ -1,0 +1,1 @@
+gcc mconsole main01.c -o main01.exe
